Добавлен сетевой запрос по обновлению аватарки профиля
