+++ ImageFeed/Profile/ProfileImageResult.swift
@@ -1,0 +1,14 @@
+//
+//  UserProfileImageResult.swift
+//  ImageFeed
+//
+//  Created by 1111 on 11.12.2024.
+//
+
+import Foundation
+
+struct ProfileImageResult: Codable {
+    let small: String?
+    let medium: String?
+    let large: String?
+}
