+++ ImageFeed/Protocols/AlertPresenterProtocol.swift
@@ -1,0 +1,8 @@
+//
+//  AlertPresenterProtocol.swift
+//  ImageFeed
+//
+//  Created by 1111 on 11.12.2024.
+//
+
+import Foundation
